openmind-ai-server/requirements.txt
# OpenMind AI Server - Dependências
# Servidor de IA para análise de imagens de produtos (ÉVORA Connect)

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Processamento de Imagens
Pillow==11.0.0

# HTTP Requests
requests==2.32.3
httpx==0.27.2

# Validação e Schemas
pydantic==2.9.2
pydantic-settings==2.5.2

# Segurança e Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Utilitários
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# IA - OpenAI (temporário para MVP)
openai==2.6.1

# Alternativa: Ollama para modelos open-source
# ollama==0.3.0

# Rate Limiting
slowapi==0.1.9

# CORS já vem integrado no FastAPI, não precisa instalar separadamente
